+++ lib/lotus/web/pages/query_editor_page.ex
@@ -15,7 +15,7 @@ defmodule Lotus.Web.QueryEditorPage do
   alias Lotus.Web.Queries.SchemaExplorerComponent
   alias Lotus.Web.Queries.VariableSettingsComponent
   alias Lotus.Web.Queries.DropdownOptionsModal
-  alias Lotus.Web.Queries.ExportCompleteModal
+  alias Lotus.Web.ExportController
   alias Lotus.Web.Formatters.VariableOptionsFormatter, as: OptionsFormatter
 
   import Lotus.Web.Queries.EditorComponent
@@ -24,7 +24,7 @@ defmodule Lotus.Web.QueryEditorPage do
   @impl Phoenix.LiveComponent
   def render(assigns) do
     ~H"""
-    <div id="query-editor-page" phx-hook="Download" class="flex flex-col h-full overflow-y-auto sm:overflow-hidden">
+    <div id="query-editor-page" class="flex flex-col h-full overflow-y-auto sm:overflow-hidden">
       <div class="mx-auto w-full px-0 sm:px-0 lg:px-6 py-0 sm:py-6 min-h-full sm:h-full flex flex-col">
         <div class="bg-white dark:bg-gray-800 shadow rounded-lg min-h-full sm:h-full flex flex-col sm:overflow-hidden">
           <.header statement_empty={@statement_empty} query={@query} mode={@page.mode} />
@@ -79,7 +79,7 @@ defmodule Lotus.Web.QueryEditorPage do
                 resolved_variable_options={@resolved_variable_options}
               />
               <div class="flex-1 overflow-y-auto sm:overflow-y-auto min-h-0">
-                <.render_result result={@result} error={@error} running={@running} os={Map.get(assigns, :os, :unknown)} target={@myself} />
+                <.render_result result={@result} error={@error} running={@running} os={Map.get(assigns, :os, :unknown)} target={@myself} is_saved_query={@page.mode == :edit} />
               </div>
 
             </div>
@@ -96,17 +96,6 @@ defmodule Lotus.Web.QueryEditorPage do
           id="dropdown_options_modal"
           variable_name={@dropdown_options_variable_name}
           variable_data={get_variable_data(@query_form, @dropdown_options_variable_name)}
-          parent={@myself}
-        />
-      <% end %>
-
-      <%= if @export_modal_visible do %>
-        <.live_component
-          module={ExportCompleteModal}
-          id="export_complete_modal"
-          filename={@export_filename}
-          file_path={@export_file_path}
-          error={@export_error}
           parent={@myself}
         />
       <% end %>
@@ -473,52 +462,8 @@ defmodule Lotus.Web.QueryEditorPage do
         {:noreply, socket}
 
       _result ->
-        {:noreply, start_streaming_export(socket)}
+        {:noreply, generate_export_url(socket)}
     end
-  end
-
-  @impl Phoenix.LiveComponent
-  def handle_event("close_export_modal", _params, socket) do
-    {:noreply,
-     assign(socket,
-       export_modal_visible: false,
-       export_filename: nil,
-       export_file_path: nil,
-       export_error: nil
-     )}
-  end
-
-  @impl Phoenix.LiveComponent
-  def handle_event("reveal_in_finder", %{"path" => file_path}, socket) do
-    case :os.type() do
-      {:unix, :darwin} ->
-        System.cmd("open", ["-R", file_path])
-
-      {:win32, _} ->
-        System.cmd("explorer", ["/select,", file_path])
-
-      _ ->
-        dir = Path.dirname(file_path)
-        System.cmd("xdg-open", [dir])
-    end
-
-    {:noreply, socket}
-  end
-
-  @impl Phoenix.LiveComponent
-  def handle_event("open_file", %{"path" => file_path}, socket) do
-    case :os.type() do
-      {:unix, :darwin} ->
-        System.cmd("open", [file_path])
-
-      {:win32, _} ->
-        System.cmd("cmd", ["/c", "start", "", file_path])
-
-      _ ->
-        System.cmd("xdg-open", [file_path])
-    end
-
-    {:noreply, socket}
   end
 
   @impl Phoenix.LiveComponent
@@ -624,20 +569,6 @@ defmodule Lotus.Web.QueryEditorPage do
     {:noreply, socket}
   end
 
-  def handle_info({:trigger_download, file_path}, socket) do
-    case File.read(file_path) do
-      {:ok, content} ->
-        {:noreply,
-         push_event(socket, "download-csv", %{
-           data: content,
-           filename: socket.assigns.export_filename
-         })}
-
-      {:error, _} ->
-        {:noreply, socket}
-    end
-  end
-
   def handle_info(_msg, socket), do: {:noreply, socket}
 
   @impl Phoenix.LiveComponent
@@ -651,37 +582,6 @@ defmodule Lotus.Web.QueryEditorPage do
 
   def handle_async(:query_execution, {:exit, _reason}, socket) do
     {:noreply, assign(socket, error: "Query execution failed", result: nil, running: false)}
-  end
-
-  @impl Phoenix.LiveComponent
-  def handle_async(:csv_export, {:ok, {:ok, file_path}}, socket) do
-    filename = Path.basename(file_path)
-
-    {:noreply,
-     assign(socket,
-       export_modal_visible: true,
-       export_filename: filename,
-       export_file_path: file_path,
-       export_task: nil
-     )}
-  end
-
-  def handle_async(:csv_export, {:ok, {:error, error}}, socket) do
-    {:noreply,
-     assign(socket,
-       export_error: to_string(error),
-       export_task: nil,
-       export_modal_visible: true
-     )}
-  end
-
-  def handle_async(:csv_export, {:exit, _reason}, socket) do
-    {:noreply,
-     assign(socket,
-       export_error: "Export failed unexpectedly",
-       export_task: nil,
-       export_modal_visible: true
-     )}
   end
 
   @impl Phoenix.LiveComponent
@@ -751,12 +651,7 @@ defmodule Lotus.Web.QueryEditorPage do
       editor_dialect: nil,
       detected_variables: [],
       variable_form: to_form(%{}, as: "variables"),
-      resolved_variable_options: %{},
-      export_modal_visible: false,
-      export_filename: nil,
-      export_file_path: nil,
-      export_error: nil,
-      export_task: nil
+      resolved_variable_options: %{}
     )
   end
 
@@ -1169,7 +1064,7 @@ defmodule Lotus.Web.QueryEditorPage do
     end
   end
 
-  defp start_streaming_export(socket) do
+  defp generate_export_url(socket) do
     query = socket.assigns.query
     vars = Map.get(socket.assigns, :variable_values, %{})
     repo = query.data_repo || socket.assigns.default_repo
@@ -1194,56 +1089,37 @@ defmodule Lotus.Web.QueryEditorPage do
 
     filename = "#{timestamp}_#{base_name}.csv"
 
-    temp_dir = System.tmp_dir!()
-    file_path = Path.join(temp_dir, filename)
+    export_params =
+      case socket.assigns.page do
+        %{mode: :edit, id: id} ->
+          %{
+            "query_id" => id,
+            "repo" => repo,
+            "vars" => vars,
+            "search_path" => query.search_path,
+            "filename" => filename
+          }
 
-    # Don't show modal yet - will show when complete
-    socket =
-      assign(socket,
-        export_modal_visible: false,
-        export_filename: filename,
-        export_file_path: file_path,
-        export_error: nil
-      )
-
-    start_async(socket, :csv_export, fn ->
-      perform_streaming_export(query, repo, vars, file_path, self())
-    end)
-  end
-
-  defp perform_streaming_export(query, repo, vars, file_path, parent_pid) do
-    opts = [
-      repo: repo,
-      vars: vars,
-      page_size: @default_page_size
-    ]
-
-    opts =
-      if query.search_path && String.trim(query.search_path) != "" do
-        Keyword.put(opts, :search_path, query.search_path)
-      else
-        opts
+        %{mode: :new} ->
+          %{
+            "query_attrs" => %{
+              "statement" => query.statement,
+              "variables" => Enum.map(query.variables || [], &variable_to_params/1)
+            },
+            "repo" => repo,
+            "vars" => vars,
+            "search_path" => query.search_path,
+            "filename" => filename
+          }
       end
 
-    try do
-      {:ok, file} = File.open(file_path, [:write, :utf8])
+    endpoint = socket.endpoint
+    token = ExportController.generate_token(endpoint, export_params)
 
-      stream = Lotus.Export.stream_csv(query, opts)
+    prefix = socket.assigns[:prefix] || ""
+    export_url = "#{prefix}/lotus/export/csv?token=#{URI.encode_www_form(token)}"
 
-      stream
-      |> Enum.each(fn chunk ->
-        IO.write(file, chunk)
-      end)
-
-      File.close(file)
-
-      send(parent_pid, {:trigger_download, file_path})
-
-      {:ok, file_path}
-    rescue
-      error ->
-        {:error, Exception.message(error)}
-    end
+    push_event(socket, "download-url", %{url: export_url})
   end
 
   defp delete_query(socket) do